pyproject.toml
[project]
name = "pdm-mypyc"
description = "A PDM build hook to compile source files with mypyc"
authors = [
    {name = "Frost Ming", email = "me@frostming.com"},
]
dependencies = [
    "mypy",
    "setuptools",
    "pdm-backend"
]
requires-python = ">=3.7"
readme = "README.md"
license = {text = "MIT"}
dynamic = ["version"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3 :: Only",
]

[project.urls]
Homepage = "https://github.com/frostming/pdm-mypyc"
Changelog = "https://github.com/frostming/pdm-mypyc/releases"

[project.entry-points."pdm.build.hook"]
mypyc = "pdm_mypyc:MypycBuildHook"

[tool.pdm]
version = {source = "scm"}

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.1.2",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
